+++ src/BLL/Widget/Widget.ts
