+++ src/components/Dashboard/MainDashboard.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script setup lang="ts">
-import DbWidget from "@/components/CoWidget.vue";
+import DbWidget from "@/components/Widget/DefaultWidget.vue";
 import { computed } from "vue";
 import DbHeader from "@/components/CoHeader.vue";
 import { useDashboard } from "@/composables/useDashboard";
