+++ src/index.ts
@@ -1,7 +1,7 @@
-import DashboardCollection from "@/BLL/DashboardCollection";
-import Dashboard from "@/BLL/Dashboard";
-import Widget from "@/BLL/Widget";
-import "./style.css";
+import DashboardCollection from "@/BLL/Dashboard/DashboardCollection";
+import Dashboard from "@/BLL/Dashboard/Dashboard";
+import Widget from "@/BLL/Widget/Widget";
+import "./Style/style.css";
 
 import DashboardDALMemoryContext from "@/DAL/DashboardDALMemoryContext";
 import WidgetDALMemoryContext from "@/DAL/WidgetDALMemoryContext";
@@ -11,6 +11,6 @@ DashboardCollection.injectDAL(new DashboardDALMemoryContext());
 Dashboard.injectDAL(new DashboardDALMemoryContext());
 Widget.injectDAL(new WidgetDALMemoryContext());
 
-import DbDashboard from "./components/CoDashboard.vue";
+import DbDashboard from "./components/Dashboard/MainDashboard.vue";
 
 export { DbDashboard };
